prettifying

+++ def.h
@@ -7,5 +7,5 @@
 
 
     typedef uint8_t u8;
-    typedef uint8_t u16;
+    typedef uint16_t u16;
 #endif
